+++ config/test.exs
@@ -1,0 +1,4 @@
+use Mix.Config
+
+config :plug_rest,
+  known_methods: ["GET", "HEAD", "OPTIONS", "POST", "MOVE"]