+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.4.0-dev
+
+* Enhancements
+  * The `resource` macro is tested to work with any Plug module
+  * The `resource` macro will accept the same options as Phoenix
+    Router's `match` macro
+
+* Backwards incompatible changes
+  * A Resource plug's options are passed directly as the `state`. In
+    resource routes, change any `state: true` to just `true`.
+  * `known_methods` defaults can be changed in the application config,
+    and are no longer an option for the Router or Resource plugs
+
 ## v0.3.1
 
 * Bug fixes